e_commerce/requirements/base.txt
# Django 
Django>=4.1,<4.3
djvue==0.2.1
sentry-sdk==1.11.1
elastic-apm==6.13.2
# for windows please add following lib before installing compressor
# rcssmin==1.1.0 --install-option="--without-c-extensions"
django-compressor==4.1
django-redis==5.2.0
redis==4.3.5

djangorestframework==3.14.0
django-environ==0.9.0
django-debug-toolbar==4.2.0
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3.1
